requirements.txt
Flask==1.1.2
mysql-connector-python==8.1.0
requests==2.26.0
selenium==4.4.3
beautifulsoup4==4.10.0